requirements.txt
anthropic
pdfplumber
python-telegram-bot==20.7
gspread
google-auth
requests
python-dotenv